Update ul_amd.js
Not sure how this could work as is...  The callback on line 81 wants a single argument... and then it again tries to `apply` it... I could only get this to work if I changed this to a `call` (as I have done in this PR) or if I changed line 81 to use the new `...args` syntax so that it accepted the FULL list of arguments trying to be passed here and then turned around and passed them on.

Without any changes only the first argument makes it to line 81 and then line 82 can't possibly work properly with the apply since you only have one argument. 

Am I missing something?